requirements.txt
fastapi
uvicorn
scikit-learn
numpy
joblib
Pillow
python-multipart
